python_app/requirements.txt
requests
psycopg2-binary
kafka-python
alembic
pandas
boto3